Move AsyncFd to its own module
So we have a place to add stuff for fixed/direct fds.
